small refactor of RegistrationProvider
