+++ apps/web/pages/names/index.tsx
@@ -1,156 +1,10 @@
-import { Transition } from '@headlessui/react';
-import { FloatingENSPills } from 'apps/web/src/components/Basenames/FloatingENSPills';
-import { LearnMoreModal } from 'apps/web/src/components/Basenames/LearnMoreModal';
-import { RegistrationContext } from 'apps/web/src/components/Basenames/RegistrationContext';
-import { RegistrationForm } from 'apps/web/src/components/Basenames/RegistrationForm';
-import ShareUsernameModal from 'apps/web/src/components/Basenames/ShareUsernameModal';
-import { UsernamePill, UsernamePillVariants } from 'apps/web/src/components/Basenames/UsernamePill';
-import { UsernameProfileForm } from 'apps/web/src/components/Basenames/UsernameProfileForm';
-import {
-  UsernameSearchInput,
-  UsernameSearchInputVariant,
-} from 'apps/web/src/components/Basenames/UsernameSearchInput';
-import tempPendingAnimation from 'apps/web/src/components/Basenames/tempPendingAnimation.png';
-import { Icon } from 'apps/web/src/components/Icon/Icon';
+import RegistrationProvider from 'apps/web/src/components/Basenames/RegistrationContext';
+import RegistrationFlow from 'apps/web/src/components/Basenames/RegistrationFlow';
 import { Layout, NavigationType } from 'apps/web/src/components/Layout/Layout';
-import {
-  findFirstValidDiscount,
-  useAggregatedDiscountValidators,
-} from 'apps/web/src/hooks/useAggregatedDiscountValidators';
-import classNames from 'classnames';
 import Head from 'next/head';
-import { Fragment, ReactElement, useCallback, useEffect, useMemo, useState } from 'react';
-import { useInterval } from 'usehooks-ts';
-import { useWaitForTransactionReceipt } from 'wagmi';
-// TODO: replace appropriate backgrounds w/Lottie files
-
-export enum ClaimProgression {
-  SEARCH,
-  CLAIM,
-  PENDING,
-  SUCCESS,
-  PROFILE,
-}
-
-export enum Discount {
-  CBID = 'CBID',
-  CB1 = 'CB1',
-  COINBASE_VERIFIED_ACCOUNT = 'COINBASE_VERIFIED_ACCOUNT',
-}
-
-const SEARCH_LABEL_COPY_STRINGS = [
-  'Set up a community profile.',
-  'Connect with Based people.',
-  'Get exclusive onchain perks.',
-];
-
-const useRotatingText = (strings: string[]) => {
-  const [currentIndex, setCurrentIndex] = useState(0);
-  useInterval(() => {
-    setCurrentIndex((prevIndex) => (prevIndex + 1) % strings.length);
-  }, 3000);
-  return strings[currentIndex];
-};
-
-/*
-test addresses w/ different verifications
-  0xB18e4C959bccc8EF86D78DC297fb5efA99550d85 - cb.id 
-  0xB18e4C959bccc8EF86D78DC297fb5efA99550d85, 0xB6944B3074F40959E1166fe010a3F86B02cF2b7c- verified account
-  0x9C02E8E28D8b706F67dcf0FC7F46A9ee1f9649FA - cb1
-*/
+import { ReactElement } from 'react';
 
 export function Usernames() {
-  const { data: discounts, loading: loadingDiscounts } = useAggregatedDiscountValidators();
-  const discount = findFirstValidDiscount(discounts);
-  const [progress, setProgress] = useState<ClaimProgression>(ClaimProgression.SEARCH);
-
-  // TODO: Not a big fan of this, I think ideally we'd have useRegisterNameCallback here
-  const [registerNameTransactionHash, setRegisterNameTransactionHash] = useState<
-    `0x${string}` | undefined
-  >();
-  const [learnMoreModalOpen, setLearnMoreModalOpen] = useState(false);
-  const toggleLearnMoreModal = useCallback(() => setLearnMoreModalOpen((open) => !open), []);
-  const [shareUsernameModalOpen, setShareUsernameModalOpen] = useState(false);
-  const toggleShareUsernameModal = useCallback(
-    () => setShareUsernameModalOpen((open) => !open),
-    [],
-  );
-
-  const [inputFocused, setInputFocused] = useState(false);
-  const [inputHovered, setInputHovered] = useState(false);
-  const [selectedName, setSelectedName] = useState('');
-
-  const onFocusLargeUsernameSearchInput = useCallback(() => setInputFocused(true), []);
-  const onBlurLargeUsernameSearchInput = useCallback(() => setInputFocused(false), []);
-  const onMouseEnterLargeUsernameSearchInput = useCallback(() => setInputHovered(true), []);
-  const onMouseLeaveLargeUsernameSearchInput = useCallback(() => setInputHovered(false), []);
-
-  const selectName = useCallback((name: string) => {
-    setProgress(ClaimProgression.CLAIM);
-    setInputFocused(false);
-    setSelectedName(name);
-  }, []);
-
-  const registrationFormOnApproved = useCallback((transactionHash: `0x${string}`) => {
-    setRegisterNameTransactionHash(transactionHash);
-  }, []);
-
-  // Wait for text record transaction to be processed
-
-  const { isFetching: transactionIsFetching, isSuccess: transactionIsSuccess } =
-    useWaitForTransactionReceipt({
-      hash: registerNameTransactionHash,
-      query: {
-        enabled: !!registerNameTransactionHash,
-      },
-    });
-
-  useEffect(() => {
-    if (transactionIsFetching) {
-      setProgress(ClaimProgression.PENDING);
-    }
-    if (transactionIsSuccess) {
-      setProgress(ClaimProgression.SUCCESS);
-    }
-  }, [transactionIsFetching, transactionIsSuccess]);
-
-  const rotatingText = useRotatingText(SEARCH_LABEL_COPY_STRINGS);
-
-  const isSearch = progress === ClaimProgression.SEARCH;
-  const isClaim = progress === ClaimProgression.CLAIM;
-  const isPending = progress === ClaimProgression.PENDING;
-  const isSuccess = progress === ClaimProgression.SUCCESS;
-  const isProfile = progress === ClaimProgression.PROFILE;
-
-  const transitionDuration = 'duration-700';
-
-  const blueBackground = inputFocused || isSuccess;
-
-  const mainClasses = classNames(
-    'relative z-10 flex min-h-screen w-full overflow-hidden flex-col items-center px-6',
-    'transition-all',
-    transitionDuration,
-    {
-      'bg-ocsblue text-white': blueBackground,
-      'bg-white text-black': !blueBackground,
-      'pt-[calc(50vh-12rem)]': isSearch || isClaim || isPending || isSuccess,
-      'pt-0': isProfile,
-    },
-  );
-
-  const claimBackgroundClasses = classNames(
-    'pointer-events-none absolute inset-0 w-full h-full bg-cover bg-center -z-10',
-  );
-
-  const registrationValue = useMemo(
-    () => ({ focused: inputFocused, hovered: inputHovered }),
-    [inputFocused, inputHovered],
-  );
-
-  const currentUsernamePillVariant = isProfile
-    ? UsernamePillVariants.Card
-    : UsernamePillVariants.Inline;
-
   return (
     <>
       <Head>
@@ -160,247 +14,9 @@ export function Usernames() {
           name="description"
         />
       </Head>
-      <RegistrationContext.Provider value={registrationValue}>
-        <main className={mainClasses}>
-          {/* TODO: REMOVE ME WHEN DONE TESTING */}
-          <div className="absolute right-20 top-40 z-50 w-[10rem] rounded-lg border border-line/20 bg-white p-4 text-black shadow-lg">
-            <ul className="flex flex-col gap-2">
-              <li>
-                <button
-                  type="button"
-                  // eslint-disable-next-line react-perf/jsx-no-new-function-as-prop
-                  onClick={() => setProgress(ClaimProgression.SEARCH)}
-                  className="rounded border border-line/10 p-2"
-                >
-                  Search
-                </button>
-              </li>
-              <li>
-                <button
-                  type="button"
-                  // eslint-disable-next-line react-perf/jsx-no-new-function-as-prop
-                  onClick={() => setProgress(ClaimProgression.CLAIM)}
-                  className="rounded border border-line/10 p-2"
-                >
-                  Claim
-                </button>
-              </li>
-              <li>
-                <button
-                  type="button"
-                  // eslint-disable-next-line react-perf/jsx-no-new-function-as-prop
-                  onClick={() => setProgress(ClaimProgression.PENDING)}
-                  className="rounded border border-line/10 p-2"
-                >
-                  Pending
-                </button>
-              </li>
-              <li>
-                <button
-                  type="button"
-                  // eslint-disable-next-line react-perf/jsx-no-new-function-as-prop
-                  onClick={() => setProgress(ClaimProgression.SUCCESS)}
-                  className="rounded border border-line/10 p-2"
-                >
-                  Success
-                </button>
-              </li>
-              <li>
-                <button
-                  type="button"
-                  // eslint-disable-next-line react-perf/jsx-no-new-function-as-prop
-                  onClick={() => setProgress(ClaimProgression.PROFILE)}
-                  className="rounded border border-line/10 p-2"
-                >
-                  Profile
-                </button>
-              </li>
-            </ul>
-          </div>
-          <Transition
-            appear
-            show={isSearch}
-            className={classNames('transition-opacity', transitionDuration)}
-            enterFrom={classNames('opacity-0')}
-            enterTo={classNames('opacity-100')}
-            leaveFrom="opacity-100"
-            leaveTo="opacity-0"
-          >
-            <FloatingENSPills />
-          </Transition>
-          <Transition
-            appear
-            show={isClaim || isPending}
-            className={classNames('transition-opacity', transitionDuration)}
-            enterFrom={classNames('opacity-0')}
-            enterTo={classNames('opacity-100')}
-            leaveFrom="opacity-100"
-            leaveTo="opacity-0"
-          >
-            {/* TODO: Lottie animation file */}
-            <div
-              className={claimBackgroundClasses}
-              style={{ backgroundImage: `url(${tempPendingAnimation.src})` }}
-            />
-          </Transition>
-          <div className="relative mx-auto mb-12 mt-24 w-full max-w-[36rem]">
-            <Transition
-              appear
-              show={isSearch}
-              className={classNames(
-                'absolute flex w-full items-center justify-between transition-opacity',
-                transitionDuration,
-              )}
-              enterFrom="opacity-0"
-              enterTo="opacity-100"
-              leaveFrom="opacity-100"
-              leaveTo="opacity-0"
-            >
-              <div className="flex items-center gap-1">
-                <Icon name="blueCircle" color="currentColor" width={15} height={15} />
-                <h1 className="text-md font-bold md:text-xl">Basenames</h1>
-              </div>
-
-              {SEARCH_LABEL_COPY_STRINGS.map((string) => (
-                <Transition
-                  as={Fragment}
-                  key={string}
-                  show={rotatingText === string}
-                  enter={classNames('transform transition delay-500', transitionDuration)}
-                  enterFrom="opacity-0"
-                  enterTo="opacity-100"
-                  leave={classNames('transform transition', transitionDuration)}
-                  leaveFrom="opacity-100"
-                  leaveTo="opacity-0"
-                >
-                  <p className="text-md absolute right-0 md:text-xl ">{string}</p>
-                </Transition>
-              ))}
-            </Transition>
-          </div>
-          <div className="relative w-full">
-            <Transition
-              appear
-              show={isClaim}
-              className={classNames(
-                'absolute left-1/2 z-40 mx-auto w-full max-w-[14rem] -translate-x-1/2 -translate-y-20 transition-all',
-                transitionDuration,
-              )}
-              enterFrom="opacity-0"
-              enterTo="opacity-100"
-              leaveFrom="opacity-100"
-              leaveTo="opacity-0"
-            >
-              <UsernameSearchInput
-                variant={UsernameSearchInputVariant.Small}
-                placeholder="Find another name"
-                selectName={selectName}
-              />
-            </Transition>
-            <div className="relative mb-40">
-              <Transition
-                appear
-                show={isClaim || isProfile || isPending || isSuccess}
-                className={classNames(
-                  'absolute left-1/2 top-0 z-30 mx-auto -translate-x-1/2 transition-all',
-                  transitionDuration,
-                  { 'animate-pulse': isPending },
-                )}
-                enter="overflow-hidden"
-                enterFrom={classNames('opacity-0 max-w-[5rem]')}
-                enterTo={classNames('opacity-100 max-w-full')}
-                leave="transition-all "
-                leaveFrom="opacity-100"
-                leaveTo="opacity-0"
-              >
-                <UsernamePill username={selectedName} variant={currentUsernamePillVariant} />
-                {isPending && (
-                  <p className="mt-6 text-center font-bold uppercase text-line">Registering...</p>
-                )}
-              </Transition>
-
-              <Transition
-                appear
-                show={isProfile}
-                className={classNames(
-                  'absolute left-1/2 top-0 z-30 mx-auto -translate-x-1/2 transition-all',
-                  'mt-[16rem] w-full max-w-[26rem] rounded-3xl p-8 shadow-xl',
-                  transitionDuration,
-                )}
-                enterFrom={classNames('opacity-0')}
-                enterTo={classNames('opacity-100')}
-                leave="transition-all "
-                leaveFrom="opacity-100"
-                leaveTo="opacity-0 "
-              >
-                <UsernameProfileForm />
-              </Transition>
-              <Transition
-                appear
-                show={isSearch}
-                className={classNames(
-                  'absolute left-1/2 top-0 z-20 mx-auto w-full max-w-[36rem] -translate-x-1/2 transition-all',
-                  transitionDuration,
-                )}
-                enterFrom="opacity-0"
-                enterTo="opacity-100"
-                leaveFrom="opacity-100"
-                leaveTo="opacity-0 max-w-[5rem]"
-              >
-                <UsernameSearchInput
-                  variant={UsernameSearchInputVariant.Large}
-                  placeholder="Search for a name"
-                  selectName={selectName}
-                  onFocus={onFocusLargeUsernameSearchInput}
-                  onBlur={onBlurLargeUsernameSearchInput}
-                  onMouseEnter={onMouseEnterLargeUsernameSearchInput}
-                  onMouseLeave={onMouseLeaveLargeUsernameSearchInput}
-                />
-              </Transition>
-            </div>
-            <Transition
-              appear
-              show={isClaim}
-              enter={classNames('transition-opacity', transitionDuration)}
-              enterFrom="opacity-0"
-              enterTo="opacity-100"
-              leave={classNames('transition-opacity', transitionDuration)}
-              leaveFrom="opacity-100"
-              leaveTo="opacity-0"
-            >
-              <RegistrationForm
-                name={selectedName}
-                loadingDiscounts={loadingDiscounts}
-                discount={discount}
-                toggleModal={toggleLearnMoreModal}
-                onApprove={registrationFormOnApproved}
-              />
-            </Transition>
-
-            <Transition
-              appear
-              show={isSuccess}
-              enter={classNames('transition-opacity', transitionDuration)}
-              enterFrom="opacity-0"
-              enterTo="opacity-100"
-              leave={classNames('transition-opacity', transitionDuration)}
-              leaveFrom="opacity-100"
-              leaveTo="opacity-0"
-            >
-              <p>hell ye brother</p>
-            </Transition>
-          </div>
-          <LearnMoreModal
-            learnMoreModalOpen={learnMoreModalOpen}
-            toggleModal={toggleLearnMoreModal}
-          />
-          <ShareUsernameModal
-            isOpen={shareUsernameModalOpen}
-            username="ultrabased"
-            toggleModal={toggleShareUsernameModal}
-          />
-        </main>
-      </RegistrationContext.Provider>
+      <RegistrationProvider>
+        <RegistrationFlow />
+      </RegistrationProvider>
     </>
   );
 }
