--- apps/web/src/components/Basenames/RegistrationContext.ts
@@ -1,3 +1,0 @@
-import { createContext } from 'react';
-
-export const RegistrationContext = createContext({ focused: false, hovered: false });
