+++ apps/web/src/components/Basenames/RegistrationContext.tsx
@@ -1,0 +1,83 @@
+import {
+  Dispatch,
+  ReactNode,
+  SetStateAction,
+  createContext,
+  useContext,
+  useMemo,
+  useState,
+} from 'react';
+
+export enum RegistrationSteps {
+  Search = 'search',
+  Claim = 'claim',
+  Pending = 'pending',
+  Success = 'success',
+  Profile = 'profile',
+}
+
+export type RegistrationContextProps = {
+  searchInputFocused: boolean;
+  setSearchInputFocused: Dispatch<SetStateAction<boolean>>;
+  searchInputHovered: boolean;
+  setSearchInputHovered: Dispatch<SetStateAction<boolean>>;
+  registrationStep: RegistrationSteps;
+  setRegistrationStep: Dispatch<SetStateAction<RegistrationSteps>>;
+};
+
+export const RegistrationContext = createContext<RegistrationContextProps>({
+  searchInputFocused: false,
+  searchInputHovered: false,
+  registrationStep: RegistrationSteps.Search,
+  setSearchInputFocused: function () {
+    return undefined;
+  },
+  setSearchInputHovered: function () {
+    return undefined;
+  },
+  setRegistrationStep: function () {
+    return undefined;
+  },
+});
+
+type RegistrationProviderProps = {
+  children?: ReactNode;
+};
+
+// Maybe not the best place for this
+export const registrationTransitionDuration = 'duration-700';
+
+export default function RegistrationProvider({ children }: RegistrationProviderProps) {
+  const [searchInputFocused, setSearchInputFocused] = useState<boolean>(false);
+  const [searchInputHovered, setSearchInputHovered] = useState<boolean>(false);
+  const [selectedName, setSelectedName] = useState<string>('');
+  const [registrationStep, setRegistrationStep] = useState<RegistrationSteps>(
+    RegistrationSteps.Search,
+  );
+
+  // TODO: RegisterName function callback
+  // TODO: wait for transaction
+
+  const values = useMemo(() => {
+    return {
+      searchInputFocused,
+      searchInputHovered,
+      setSearchInputFocused,
+      setSearchInputHovered,
+      selectedName,
+      setSelectedName,
+      registrationStep,
+      setRegistrationStep,
+    };
+  }, [registrationStep, searchInputFocused, searchInputHovered, selectedName]);
+
+  return <RegistrationContext.Provider value={values}>{children}</RegistrationContext.Provider>;
+}
+
+export function useRegistration() {
+  const context = useContext(RegistrationContext);
+  if (context === undefined) {
+    throw new Error('useCount must be used within a CountProvider');
+  }
+  return context;
+}
